encrypt と decrypt のファイル分割
